requirements.txt
arabic-reshaper==3.0.0
certifi==2025.1.31
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
dotenv==0.9.9
elastic-transport==8.17.0
elasticsearch==8.17.1
iniconfig==2.1.0
mock==5.2.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-bidi==0.6.6
python-dotenv==1.0.1
urllib3==2.3.0
